Moved out classes to separate file
